+++ src/core/advisors/strategies/place_advisor_strategy.rs
@@ -2,7 +2,7 @@ use crate::core::link_struct::PageLink;
 
 /// Trait that is used for place advisor strategies.
 /// Place advisor strategy is used to provide place for new data.
-/// It is using EmptyLinkRegistry to find empty places.
+/// It is using [`EmptyLinkRegistry`] to find empty places.
 pub trait PlaceAdvisorStrategy {
     /// Provides place for new data.
     /// # Arguments
@@ -10,13 +10,13 @@ pub trait PlaceAdvisorStrategy {
     /// # Returns
     /// * `Option<PageLink>` - link to place where data should be placed,
     /// None if there is no place
-    fn provide_place(&mut self, size: u64) -> Option<PageLink>;
+    fn provide_place(&mut self, size: u16) -> Option<PageLink>;
 
     /// Applies place to link.
     /// # Arguments
     /// * `link` - link to place where data should be placed
     /// * `size` - size of data that should be placed
-    fn apply_place(&mut self, link: &PageLink, size: u64);
+    fn apply_place(&mut self, link: &PageLink, size: u16);
 
     /// Returns name of strategy.
     /// # Returns
